+++ rollup.browser.config.mjs
@@ -2,7 +2,7 @@ import typescript from 'rollup-plugin-typescript2';
 import commonjs from '@rollup/plugin-commonjs';
 import resolve from '@rollup/plugin-node-resolve';
 import replace from '@rollup/plugin-replace';
-import {terser} from 'rollup-plugin-terser';
+import terser from '@rollup/plugin-terser';
 import peerDepsExternal from 'rollup-plugin-peer-deps-external';
 
 const env = process.env.NODE_ENV;
@@ -14,16 +14,20 @@ export default {
     typescript(),
     resolve({jsnext: true, main: true, browser: true}),
     commonjs(),
-    replace({'process.env.NODE_ENV': JSON.stringify(env)}),
+    replace({
+      preventAssignment: true,
+      values: {'process.env.NODE_ENV': JSON.stringify(env)}
+    }),
     env === 'production' && terser()
   ],
   output: {
     name: 'Geosuggest',
     sourcemap: false,
     globals: {
-      'google': 'google',
-      'react': 'React',
-      'react-dom': 'ReactDOM'
+      google: 'google',
+      react: 'React',
+      'react-dom': 'ReactDOM',
+      'react-dom/client': 'ReactDOM'
     },
     format: 'iife',
     file: