+++ rollup.example.config.mjs
@@ -2,22 +2,22 @@ import typescript from 'rollup-plugin-typescript2';
 import commonjs from '@rollup/plugin-commonjs';
 import resolve from '@rollup/plugin-node-resolve';
 import replace from '@rollup/plugin-replace';
-import {terser} from 'rollup-plugin-terser';
+import terser from '@rollup/plugin-terser';
 import peerDepsExternal from 'rollup-plugin-peer-deps-external';
 
 const env = process.env.NODE_ENV;
 
 export default {
   input: 'example/src/app.tsx',
-
   output: {
     name: 'Geosuggest',
     sourcemap: env === 'production' ? false : 'inline',
     exports: 'none',
     globals: {
-      'google': 'google',
-      'react': 'React',
-      'react-dom': 'ReactDOM'
+      google: 'google',
+      react: 'React',
+      'react-dom': 'ReactDOM',
+      'react-dom/client': 'ReactDOM'
     },
     format: 'iife',
     file: 'example/dist/app.js'
@@ -29,7 +29,10 @@ export default {
     commonjs({
       include: 'node_modules/**'
     }),
-    replace({'process.env.NODE_ENV': JSON.stringify(env)}),
+    replace({
+      preventAssignment: true,
+      values: {'process.env.NODE_ENV': JSON.stringify(env)}
+    }),
     env === 'production' && terser()
   ]
 };