+++ src/index.ts
@@ -1,0 +1,15 @@
+/**
+ * The "codegen" part in this file is auto-generated.
+ * Please don't edit the generated exports manually.
+ *
+ * This file is needed for correct TS declaration exports in the npm module.
+ * All files with types, that should be exported from the npm module,
+ * must be re-exported in this barrel file.
+ */
+export {default} from './Geosuggest';
+// codegen:start {preset: barrel, include: ./types/*.ts}
+export * from './types/fixture';
+export * from './types/location';
+export * from './types/props';
+export * from './types/suggest';
+// codegen:end