+++ src/suggest-item.tsx
@@ -1,20 +1,21 @@
 import * as React from 'react';
 import classnames from 'classnames';
-import ISuggest from './types/suggest';
 
-interface IProps {
+import {Suggest} from './types/suggest';
+
+interface Props {
   readonly userInput: string;
   readonly isActive: boolean;
   readonly className: string;
   readonly suggestItemClassName?: string;
   readonly activeClassName?: string;
-  readonly suggest: ISuggest;
+  readonly suggest: Suggest;
   readonly isHighlightMatch: boolean;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   readonly style: any;
-  readonly onSelect: (suggest: ISuggest) => void;
+  readonly onSelect: (suggest: Suggest) => void;
   readonly renderSuggestItem?: (
-    suggest: ISuggest,
+    suggest: Suggest,
     userInput: string
   ) => JSX.Element | string;
   readonly onMouseDown: (event: React.MouseEvent) => void;
@@ -24,7 +25,7 @@ interface IProps {
 /**
  * A single Geosuggest item in the list
  */
-export default class extends React.PureComponent<IProps, unknown> {
+export default class SuggestItem extends React.PureComponent<Props, unknown> {
   /**
    * The reference to the suggest element
    */
@@ -33,7 +34,7 @@ export default class extends React.PureComponent<IProps, unknown> {
   /**
    * The constructor.
    */
-  constructor(props: IProps) {
+  constructor(props: Props) {
     super(props);
 
     this.onClick = this.onClick.bind(this);
@@ -53,10 +54,7 @@ export default class extends React.PureComponent<IProps, unknown> {
   /**
    * Replace matched text with the same in bold
    */
-  formatMatchedText(
-    userInput: string,
-    suggest: ISuggest
-  ): JSX.Element | string {
+  formatMatchedText(userInput: string, suggest: Suggest): JSX.Element | string {
     if (!userInput || !suggest.matchedSubstrings) {
       return suggest.label;
     }
@@ -91,7 +89,7 @@ export default class extends React.PureComponent<IProps, unknown> {
   /**
    * Checking if item just became active and scrolling if needed.
    */
-  componentDidUpdate(prevProps: IProps): void {
+  componentDidUpdate(prevProps: Props): void {
     if (!prevProps.isActive && this.props.isActive) {
       this.scrollIfNeeded();
     }