+++ src/suggest-list.tsx
@@ -1,16 +1,17 @@
 import * as React from 'react';
 import classnames from 'classnames';
 import SuggestItem from './suggest-item';
-import ISuggest from './types/suggest';
 
-interface IProps {
+import {Suggest} from './types/suggest';
+
+interface Props {
   readonly isHidden: boolean;
-  readonly suggests: ISuggest[];
+  readonly suggests: Suggest[];
   readonly suggestsClassName?: string;
   readonly hiddenClassName?: string;
   readonly suggestItemClassName?: string;
   readonly suggestItemActiveClassName?: string;
-  readonly activeSuggest: ISuggest | null;
+  readonly activeSuggest: Suggest | null;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   readonly style: any;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -18,12 +19,13 @@ interface IProps {
   readonly userInput: string;
   readonly isHighlightMatch: boolean;
   readonly listId: string;
+  readonly listLabel: string;
   readonly onSuggestNoResults: () => void;
   readonly renderSuggestItem?: (
-    suggest: ISuggest,
+    suggest: Suggest,
     userInput: string
   ) => JSX.Element | string;
-  readonly onSuggestSelect: (suggest: ISuggest) => void;
+  readonly onSuggestSelect: (suggest: Suggest) => void;
   readonly onSuggestMouseDown: (event: React.MouseEvent) => void;
   readonly onSuggestMouseOut: (event: React.MouseEvent) => void;
 }
@@ -31,7 +33,7 @@ interface IProps {
 /**
  * The list with suggestions.
  */
-export default class extends React.PureComponent<IProps, unknown> {
+export default class SuggestList extends React.PureComponent<Props, unknown> {
   /**
    * Whether or not it is hidden
    */
@@ -42,7 +44,7 @@ export default class extends React.PureComponent<IProps, unknown> {
   /**
    * There are new properties available for the list
    */
-  componentDidUpdate(prevProps: IProps): void {
+  componentDidUpdate(prevProps: Props): void {
     if (prevProps.suggests !== this.props.suggests) {
       if (this.props.suggests.length === 0) {
         this.props.onSuggestNoResults();
@@ -71,6 +73,7 @@ export default class extends React.PureComponent<IProps, unknown> {
         className={classes}
         style={this.props.style}
         role="listbox"
+        aria-label={this.props.listLabel}
         id={this.props.listId}>
         {this.props.suggests.map((suggest) => {
           const isActive =