+++ src/types/suggest.ts
@@ -1,9 +1,7 @@
-import IFixture from './fixture';
-
 /**
  * The suggest interface
  */
-export default interface ISuggest {
+export interface Suggest {
   readonly description?: string;
   readonly label: string;
   readonly placeId: string;